Add files via upload
Sure, here is a description for a GitHub repository on the success of movies:

Repository name: Success of Movies

Description:

This repository contains code and data for analyzing the factors that contribute to the success of movies. The code uses the The Movie Database (TMDb) API to collect data on movies, such as their genre, budget, release date, and box office gross. The data is then analyzed to identify patterns and trends that can help us understand what makes a movie successful.